+++ cli/src/commands/graph/common/fetch-schema.ts
@@ -31,7 +31,8 @@ export default (opts: CommonGraphCommandOptions) => {
       if (resp.response?.details) {
         console.log(pc.red(pc.bold(resp.response?.details)));
       }
-      process.exit(1);
+      process.exitCode = 1;
+      return;
     }
 
     if (options.out) {